refactor: move qs params into state, rename interface, move searchbar to search folder
