+++ src/pages/Home.tsx
@@ -1,6 +1,7 @@
+import { DEFAULT_PAGE_NUMBER, DEFAULT_PAGE_SIZE } from '../axios-config';
 import ErrorBoundary from '../components/ErrorBoundary';
 import SearchResults from '../components/search/SearchResults';
-import SearchBar from '../components/searchbar/SearchBar';
+import SearchBar from '../components/search/SearchBar';
 import { IState } from '../model/state.interface';
 
 function Home() {
@@ -9,6 +10,8 @@ function Home() {
     isLoading: true,
     error: '',
     items: [],
+    pageNumber: DEFAULT_PAGE_NUMBER,
+    pageSize: DEFAULT_PAGE_SIZE,
     searchTerm,
   };
